requirements.txt
Flask~=2.2.2
PyExecJS~=1.5.1
requests~=2.28.1